Add parallel feature
* Add parallel features, to (optionally) spin up rayon to decompress
  the blocks in workerthreads
