+++ backhand-cli/Cargo.toml
@@ -46,6 +46,8 @@ lzo = ["backhand/lzo"]
 ## Enables zstd compression inside library and binaries
 zstd = ["backhand/zstd"]
 lz4 = ["backhand/lz4"]
+## Enable backhand parallel decompression
+backhand-parallel = ["backhand/parallel"]
 
 [[bin]]
 name = "unsquashfs-backhand"
